+++ src/muncher/spells.js
@@ -128,6 +128,7 @@ export async function parseSpells(ids = null) {
   const itemHandler = new DDBItemImporter("spells", uniqueSpells);
   await itemHandler.init();
   await itemHandler.srdFiddling();
+  await itemHandler.iconAdditions();
   const filteredSpells = (ids !== null && ids.length > 0)
     ? itemHandler.documents.filter((s) => s.flags?.ddbimporter?.definitionId && ids.includes(String(s.flags.ddbimporter.definitionId)))
     : itemHandler.documents;